include systemvars
