--- .env.dev
@@ -1,10 +1,0 @@
-// .env
-//API_HOST=https://jobcore-test.herokuapp.com
-API_HOST=https://jobcore.herokuapp.com
-GOOGLE_MAPS_KEY=AIzaSyDsJKar6kRvBW6_y84fmT9fEJgj3F0djSY
-DEBUG=true
-PLAID_PUBLIC_KEY=037db4bdc6e10afa8721a838cc6138
-//PLAID_ENVIRONMENT=sandbox
-PLAID_ENVIRONMENT=production
-APPSTORE=https://apps.apple.com/us/app/jobcore-talent/id1437290430
-ANDROID=https://play.google.com/store/apps/details?id=co.jobcore.talent&hl=en_US
