+++ .env.example
@@ -1,7 +1,9 @@
 // .env
-API_HOST=https://8080-dcdfde83-46b3-46fb-826e-0ae656820bf6.ws-us02.gitpod.io
 GOOGLE_MAPS_KEY=sdfsfd
 DEBUG=true
+
+PLAID_PUBLIC_KEY=asdasd
+PLAID_ENVIRONMENT=production
 
 APPSTORE=sdfsdf
 ANDROID=sdfsdf
