+++ src/js/utils/validation.js
@@ -1,4 +1,5 @@
 import _validator from 'validator';
+import { snakeCase } from 'snake-case';
 
 export const validator = _validator;
 export const ValidationError = function(msg){
@@ -12,4 +13,13 @@ export function onlyLetters(value)
     if(value.match(letters)) return true;
     else return false;
 }
+
+export const normalizeToSnakeCase = (map) => {
+    const entries = Object.entries(map);
+    const newMap = {};
+    entries.forEach(([key, value]) => {
+      newMap[snakeCase(key)] = value;
+    });
+    return newMap;
+  };
