+++ webpack.dev.js
@@ -17,8 +17,5 @@ module.exports = merge(common, {
     filename: 'bundle.js',
     path: __dirname + './dist',
     publicPath: '/'
-  },
-  plugins: [
-    new Dotenv({path: '.' + path.sep + '.env.dev'})
-  ]
+  }
 });
