+++ webpack.prod.js
@@ -15,8 +15,5 @@ module.exports = merge(common, {
             }
         }
     }
-  },
-  plugins: [
-    new Dotenv({ path: './.env.prod'})
-  ]
+  }
 });
